Improve editor selection in commit message wizard
Refactor the script to allow for more flexible editor selection when
opening the commit message for review. The new logic first checks for
the GIT_EDITOR environment variable and falls back to the EDITOR
variable. If neither is set, it defaults to 'vi'. This change aims to
enhance user experience by ensuring the preferred editor is used
without reopening it unnecessarily during the commit process.

Signed-off-by: Denis Yuji Shimizu <denis.shimizu@collabora.com>